~ adding wait_random func

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -9,7 +9,7 @@ import typing
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: int) -> typing.List[int]:
+async def wait_n(n: int, max_delay: int) -> typing.List[float]:
     """
     # wait_n
     wait_n: waits up to max_delay for n times.
